e a jour

+++ Client/include/bonus.h
@@ -8,13 +8,13 @@
 class Bonus
 {
     public:
-        Bonus(sf::Vector2f position, sf::Color couleur1,sf::Color couleur2, int value);
+        Bonus(sf::Vector2f position, sf::Color couleur1,sf::Color couleur2, float value);
         virtual ~Bonus();
         void Draw(sf::RenderWindow* p_window);
         sf::Vector2f getPosition();
         sf::FloatRect getGlobalBound();
         void Move();
-        int bonusVal();
+        float bonusVal();
         sf::Color getFillColor();
         sf::Color getOutlineColor();
         sf::Vector2f getSize();
@@ -22,7 +22,7 @@ class Bonus
     protected:
         sf::RectangleShape* m_mainShape;
         sf::Vector2f m_speed;
-        int m_bonusValue;
+        float m_bonusValue;
         sf::Clock m_clock;
